refactor: ui: Extract metrics summary into modular utils
