+++ src/ui/components/mod.rs
@@ -2,6 +2,13 @@ pub mod instance_details;
 pub mod metrics_summary;
 pub mod rds_list;
 
+pub mod display_utils;
+pub mod metric_list_utils;
+pub mod metric_utils;
+pub mod sparkline_utils;
+pub mod time_range_utils;
+pub mod visual_utils;
+
 pub use instance_details::render_instance_details;
 pub use metrics_summary::render_metrics_summary;
 pub use rds_list::render_rds_list;
